pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ci_cd_tool"
version = "0.2.6"
description = "CI/CD 자동화 도구"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "click>=8.1.7",
    "rich>=13.7.0",
    "dependency-injector>=4.41.0",
    "pyyaml>=6.0.1", 
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "GitPython>=3.1.0",
    "coverage>=7.4.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "flake8>=7.0.0"
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["ci_cd_tool"]

[project.scripts]
cc = "ci_cd_tool.cli:main"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=src/ci_cd_tool"
testpaths = [
    "tests"
]